Project (#23)
* Changes of State potential done

Rodrigo Sousa <189535>

* JavaDoc Update

* Remove 2 ASK's

* Advancements on everything that surrounds Project and Submissions

* Checked for mistakes or TODO's important to mention that were forgotten

+++ sth-core/src/sth/SurveyEntry.java
@@ -5,10 +5,10 @@ import java.io.Serializable;
 public class SurveyEntry implements Serializable {
   private static final long serialVersionUID = 201810051538L;
 
-  private double _spentHours;
+  private float _spentHours;
   private String _comment;
 
-  SurveyEntry(double spentHours, String comment) {
+  SurveyEntry(float spentHours, String comment) {
     _spentHours = spentHours;
     _comment = comment;
   }
